Find the number of stomatal rows

Find the number of stomatal rows

number_rows_stomata(cells, stom.val)

Arguments

cells

A `sf` object containing all cell polygons (i.e. not just the stomatal polygons).

stom.val

Numeric. The value (colour) of stomata.

Value

A list with components:

n.rows

Numeric. The number of rows of stomata

row.bounds

Numeric. The y coordinates defining the edge of each row

row.wiggliness

Numeric. The width of the stomatal band; see Details

row.consistency

Numeric. The number of stomata per row.

stomrow.density

Numeric. The number of rows of stomata per unit length (per pixel if unscaled)

Details

This function assumes that the pixels are coloured in the following order (darkest to lightest): wall, stomate, subsidiary, pavement. To calculate row.wiggliness, the major axis of the stomatal row is determined using principal components, where PC1 is the main axis of the row and PC2 is the vertical offset of each stomata from this axis. The value row.wiggliness is the proportion of variation explained by PC2; a value close to 0 means that stomata are perfectly aligned; a value close to 0.5 means that stomatal rows are variable or wiggly in alignment. Values cannot exceed 0.5.

@import stats